added logic for jammy on python cryptography pkg

+++ scripts/prepare-env.sh
@@ -2116,7 +2116,10 @@ chmod 0755 ~/.local/bin/*
 pylxd=$(pip3 list 2> /dev/null | grep pylxd)
 if [ "$pylxd" == "" ]; then
   pip3 install pylxd 2> /dev/null
-  pip3 uninstall -y cryptography 2> /dev/null
+  codename=$(lsb_release -a 2> /dev/null | grep Codename | awk '{print $2}')
+  if [ "$codename" != "jammy"]; then
+    pip3 uninstall -y cryptography 2> /dev/null
+  fi
 fi
 
 lxdg=$(id | sed 's/^.*\(lxd\).*$/\1/')
